requirements.txt
osmnx>=1.3.0
networkx>=3.1
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.10.0


folium>=0.14.0
Pillow>=10.0.0
